Fixed: crash at Guild::AddMember based on bad data type, patch provided by mirek
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@3659 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Player.cpp
@@ -4693,8 +4693,8 @@ uint32 Player::GetZoneIdFromDB(uint64 guid)
     Field* fields = result->Fetch();
 
     uint32 map  = fields[0].GetUInt32();
-    uint32 posx = fields[1].GetFloat();
-    uint32 posy = fields[2].GetFloat();
+    float posx = fields[1].GetFloat();
+    float posy = fields[2].GetFloat();
 
     delete result;
